#269 Updated streaming documentation
