+++ src/client/src/components/form/requestForms/BroadcastingForm.vue
@@ -25,8 +25,7 @@
 import FormArea from '@/components/form/FormArea.vue'
 import IftaFormField from '@/components/form/IftaFormField.vue'
 import { Button, InputText, useToast } from 'primevue'
-import { type Reactive, reactive, ref } from 'vue'
-import type { ConnectionInformationResponse } from '@/api/resources/streamApi.ts'
+import { onMounted, type Ref, ref } from 'vue'
 import { streamerRequestRules } from '@/validation/requestRules.ts'
 import { createFormValidation } from '@/validation/types/formValidation.ts'
 import usersApi, { type StreamerRequest } from '@/api/resources/usersApi.ts'
@@ -35,18 +34,21 @@ import { useConnectionInfoStore } from '@/stores/connectionInfoStore.ts'
 import { showEditSuccessToast } from '@/utils/toastUtils.ts'
 
 const toast = useToast()
-const props = defineProps<{
-  info: ConnectionInformationResponse
-}>()
+
 const connectionInfoStore = useConnectionInfoStore()
-const formState: Reactive<StreamerRequest> = reactive({ displayName: props.info.displayName })
+const formState: Ref<StreamerRequest> = ref({ displayName: '' })
 const validation = createFormValidation(formState, streamerRequestRules)
 const isSubmitting = ref(false)
+
+onMounted(async () => {
+  await connectionInfoStore.loadIfNotInitialized()
+  formState.value.displayName = connectionInfoStore.liveInfo()?.displayName ?? ''
+})
 
 const handleSave = async () => {
   if (!validation.validate()) return
   isSubmitting.value = true
-  const updateResponse = await usersApi.putStreamer(formState)
+  const updateResponse = await usersApi.putStreamer(formState.value)
   if (tryHandleUnsuccessfulResponse(updateResponse, toast, validation)) {
     isSubmitting.value = false
     return
